Add fp32 versions
