+++ runtime/nvqir/cutensornet/simulator_mps_fp64_register.cpp
@@ -1,0 +1,11 @@
+/*******************************************************************************
+ * Copyright (c) 2022 - 2025 NVIDIA Corporation & Affiliates.                  *
+ * All rights reserved.                                                        *
+ *                                                                             *
+ * This source code and the accompanying materials are made available under    *
+ * the terms of the Apache License 2.0 which accompanies this distribution.    *
+ ******************************************************************************/
+
+#include "simulator_mps.h"
+
+NVQIR_REGISTER_SIMULATOR(nvqir::SimulatorMPS<double>, tensornet_mps)
