+++ runtime/nvqir/cutensornet/simulator_tensornet.h
@@ -1,4 +1,4 @@
-/*******************************************************************************
+/****************************************************************-*- C++ -*-****
  * Copyright (c) 2022 - 2025 NVIDIA Corporation & Affiliates.                  *
  * All rights reserved.                                                        *
  *                                                                             *
@@ -6,12 +6,17 @@
  * the terms of the Apache License 2.0 which accompanies this distribution.    *
  ******************************************************************************/
 
+#pragma once
 #include "cudaq.h"
 #include "simulator_cutensornet.h"
 #include "tn_simulation_state.h"
 
 // Forward declaration
+#ifdef TENSORNET_FP32
+extern "C" nvqir::CircuitSimulator *getCircuitSimulator_tensornet_fp32();
+#else
 extern "C" nvqir::CircuitSimulator *getCircuitSimulator_tensornet();
+#endif
 
 namespace nvqir {
 template <typename ScalarType = double>
@@ -57,7 +62,11 @@ public:
 
   // Nothing to do for state preparation
   virtual void prepareQubitTensorState() override {}
+#ifdef TENSORNET_FP32
+  virtual std::string name() const override { return "tensornet-fp32"; }
+#else
   virtual std::string name() const override { return "tensornet"; }
+#endif
   CircuitSimulator *clone() override {
     thread_local static auto simulator = std::make_unique<SimulatorTensorNet>();
     return simulator.get();
@@ -150,27 +159,12 @@ public:
   }
 
 private:
+#ifdef TENSORNET_FP32
+  friend nvqir::CircuitSimulator * ::getCircuitSimulator_tensornet_fp32();
+#else
   friend nvqir::CircuitSimulator * ::getCircuitSimulator_tensornet();
+#endif
   /// @brief Has cuTensorNet MPI been initialized?
   bool m_cutnMpiInitialized = false;
 };
 } // namespace nvqir
-
-/// Register this Simulator class with NVQIR under name "tensornet"
-extern "C" {
-nvqir::CircuitSimulator *getCircuitSimulator_tensornet() {
-  thread_local static auto simulator =
-      std::make_unique<nvqir::SimulatorTensorNet<float>>();
-  // Handle multiple runtime __nvqir__setCircuitSimulator calls before/after MPI
-  // initialization. If the static simulator instance was created before MPI
-  // initialization, it needs to be reset to support MPI if needed.
-  if (cudaq::mpi::is_initialized() && !simulator->m_cutnMpiInitialized) {
-    // Reset the static instance to pick up MPI.
-    simulator.reset(new nvqir::SimulatorTensorNet<float>());
-  }
-  return simulator.get();
-}
-nvqir::CircuitSimulator *getCircuitSimulator() {
-  return getCircuitSimulator_tensornet();
-}
-}
