+++ runtime/nvqir/cutensornet/simulator_tensornet_fp64_register.cpp
@@ -1,0 +1,28 @@
+/*******************************************************************************
+ * Copyright (c) 2022 - 2025 NVIDIA Corporation & Affiliates.                  *
+ * All rights reserved.                                                        *
+ *                                                                             *
+ * This source code and the accompanying materials are made available under    *
+ * the terms of the Apache License 2.0 which accompanies this distribution.    *
+ ******************************************************************************/
+
+#include "simulator_tensornet.h"
+
+/// Register this Simulator class with NVQIR under name "tensornet"
+extern "C" {
+nvqir::CircuitSimulator *getCircuitSimulator_tensornet() {
+  thread_local static auto simulator =
+      std::make_unique<nvqir::SimulatorTensorNet<double>>();
+  // Handle multiple runtime __nvqir__setCircuitSimulator calls before/after MPI
+  // initialization. If the static simulator instance was created before MPI
+  // initialization, it needs to be reset to support MPI if needed.
+  if (cudaq::mpi::is_initialized() && !simulator->m_cutnMpiInitialized) {
+    // Reset the static instance to pick up MPI.
+    simulator.reset(new nvqir::SimulatorTensorNet<double>());
+  }
+  return simulator.get();
+}
+nvqir::CircuitSimulator *getCircuitSimulator() {
+  return getCircuitSimulator_tensornet();
+}
+}
