+++ unittests/CMakeLists.txt
@@ -92,6 +92,14 @@ macro (create_tests_with_backend NVQIR_BACKEND EXTRA_BACKEND_TESTER)
     target_compile_definitions(${TEST_EXE_NAME} PRIVATE -DCUDAQ_BACKEND_TENSORNET -DCUDAQ_BACKEND_TENSORNET_MPS -DCUDAQ_SIMULATION_SCALAR_FP64)
     set(TEST_LABELS "gpu_required")
   endif()
+  if (${NVQIR_BACKEND} STREQUAL "tensornet-fp32")
+    target_compile_definitions(${TEST_EXE_NAME} PRIVATE -DCUDAQ_BACKEND_TENSORNET -DCUDAQ_SIMULATION_SCALAR_FP32)
+    set(TEST_LABELS "gpu_required")
+  endif()
+  if (${NVQIR_BACKEND} STREQUAL "tensornet-mps-fp32")
+    target_compile_definitions(${TEST_EXE_NAME} PRIVATE -DCUDAQ_BACKEND_TENSORNET -DCUDAQ_BACKEND_TENSORNET_MPS -DCUDAQ_SIMULATION_SCALAR_FP32)
+    set(TEST_LABELS "gpu_required")
+  endif()
   if (${NVQIR_BACKEND} STREQUAL "custatevec-fp32")
     target_compile_definitions(${TEST_EXE_NAME} PRIVATE -DCUDAQ_BACKEND_CUSTATEVEC_FP32 -DCUDAQ_SIMULATION_SCALAR_FP32)
     target_link_libraries(${TEST_EXE_NAME} PRIVATE ${CUDA_LIBRARIES} ${CUDA_CUDART_LIBRARY})
@@ -200,6 +208,8 @@ if(TARGET nvqir-tensornet)
   message(STATUS "Building cutensornet backend tests.")
   create_tests_with_backend(tensornet "")
   create_tests_with_backend(tensornet-mps "")
+  create_tests_with_backend(tensornet-fp32 "")
+  create_tests_with_backend(tensornet-mps-fp32 "")
   if (MPI_CXX_FOUND)
     # Count the number of GPUs
     find_program(NVIDIA_SMI "nvidia-smi")
